+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 - Fix false positive in `ReturnTypeDeclarationSniff` with nullable types.
+- Relax check for missing return type when `{aType}|null` doc bloc is present.
 - Add `is` to the list of allowed short names.
 
 ## 0.10.0